implemented pop method in singly linked list
In this, we need to take two variables one will be pointing toward the start of the element and the second will be pointing towards the next element of the list. next, we need to traverse the second variable value to the last of the list. next, we need to cut the link of the last element and make the 2nd last element as the last element..